+++ actionview/lib/action_view/render_parser.rb
@@ -6,47 +6,73 @@ module ActionView
   class RenderParser # :nodoc:
     ALL_KNOWN_KEYS = [:partial, :template, :layout, :formats, :locals, :object, :collection, :as, :status, :content_type, :location, :spacer_template]
     RENDER_TYPE_KEYS = [:partial, :template, :layout]
+    RENDER_CALL_METHODS = /\A(render|render_to_string)\z/
+    LAYOUT_CALL_METHOD = /\Alayout\z/
 
-    def initialize(name, code)
+    RenderCallInfo = Struct.new(:virtual_path, :locals_keys)
+
+    def initialize(name, code, from_controller: false)
       @name = name
       @code = code
+      @from_controller = from_controller
     end
 
     def render_calls
-      queue = [Prism.parse(@code).value]
-      templates = []
+      collect_render_calls.map(&:virtual_path)
+    end
 
-      while (node = queue.shift)
-        queue.concat(node.compact_child_nodes)
-        next unless node.is_a?(Prism::CallNode)
-
-        options = render_call_options(node)
-        next unless options
-
-        render_type = (options.keys & RENDER_TYPE_KEYS)[0]
-        template, object_template = render_call_template(options[render_type])
-        next unless template
-
-        if options.key?(:object) || options.key?(:collection) || object_template
-          next if options.key?(:object) && options.key?(:collection)
-          next unless options.key?(:partial)
-        end
-
-        if options[:spacer_template].is_a?(Prism::StringNode)
-          templates << partial_to_virtual_path(:partial, options[:spacer_template].unescaped)
-        end
-
-        templates << partial_to_virtual_path(render_type, template)
-
-        if render_type != :layout && options[:layout].is_a?(Prism::StringNode)
-          templates << partial_to_virtual_path(:layout, options[:layout].unescaped)
-        end
-      end
-
-      templates
+    def render_calls_with_locals
+      collect_render_calls
     end
 
     private
+      def collect_render_calls
+        queue = [Prism.parse(@code).value]
+        results = []
+
+        while (node = queue.shift)
+          queue.concat(node.compact_child_nodes)
+          next unless node.is_a?(Prism::CallNode)
+
+          if @from_controller && node.name.to_s.match?(LAYOUT_CALL_METHOD) && !node.receiver && node.arguments
+            call = parse_layout_call(node)
+            results << call if call
+            next
+          end
+
+          options = render_call_options(node)
+          next unless options
+
+          render_type = (options.keys & RENDER_TYPE_KEYS)[0]
+          template, object_template = render_call_template(options[render_type])
+          next unless template
+
+          locals_keys = extract_locals_keys(options, render_type, template, object_template)
+          next unless locals_keys
+
+          if options[:spacer_template].is_a?(Prism::StringNode)
+            results << RenderCallInfo.new(
+              partial_to_virtual_path(:partial, options[:spacer_template].unescaped),
+              locals_keys.dup
+            )
+          end
+
+          results << RenderCallInfo.new(
+            partial_to_virtual_path(render_type, template),
+            locals_keys
+          )
+
+          if render_type != :layout && options[:layout].is_a?(Prism::StringNode)
+            results << RenderCallInfo.new(
+              partial_to_virtual_path(:layout, options[:layout].unescaped),
+              locals_keys.dup
+            )
+          end
+        end
+
+        results
+      end
+
       def directory
         File.dirname(@name)
       end
@@ -57,6 +83,63 @@ module ActionView
         else
           partial_path
         end
+      end
+
+      def extract_locals_keys(options, render_type, template, object_template)
+        locals_keys = []
+
+        if options.key?(:locals)
+          locals_node = options[:locals]
+          if locals_node.is_a?(Prism::HashNode) || locals_node.is_a?(Prism::KeywordHashNode)
+            if locals_node.elements.all? { |e| e.is_a?(Prism::AssocNode) && e.key.is_a?(Prism::SymbolNode) }
+              locals_keys = locals_node.elements.map { |e| e.key.unescaped.to_sym }
+            else
+              return nil
+            end
+          elsif !locals_node.nil?
+            return nil
+          end
+        end
+
+        if options.key?(:object) || options.key?(:collection) || object_template
+          return nil if options.key?(:object) && options.key?(:collection)
+          return nil unless options.key?(:partial)
+
+          as = if options.key?(:as)
+            parse_symbol_or_string(options[:as])
+          else
+            File.basename(template)[/\A_?(.*?)(?:\.\w+)*\z/, 1]
+          end
+          return nil unless as
+
+          locals_keys << as.to_sym
+          if options.key?(:collection)
+            locals_keys << :"#{as}_counter"
+            locals_keys << :"#{as}_iteration"
+          end
+        end
+
+        locals_keys.sort!
+        locals_keys
+      end
+
+      def parse_symbol_or_string(node)
+        if node.is_a?(Prism::SymbolNode)
+          node.unescaped
+        elsif node.is_a?(Prism::StringNode)
+          node.unescaped
+        end
+      end
+
+      def parse_layout_call(node)
+        arguments = node.arguments.arguments
+        return unless arguments.length == 1
+
+        arg = arguments[0]
+        return unless arg.is_a?(Prism::StringNode)
+
+        virtual_path = "layouts/#{arg.unescaped}"
+        RenderCallInfo.new(virtual_path, [])
       end
 
       # Accept a call node and return a hash of options for the render call.
@@ -90,7 +173,12 @@ module ActionView
         # array of locals or a keyword hash with symbol keys.
         options =
           if (length == 1 || length == 2) && !arguments[0].is_a?(Prism::KeywordHashNode)
-            { partial: arguments[0], locals: arguments[1] }
+            if @from_controller
+              # In controller context, render("foo") means render(template: "foo")
+              { template: arguments[0], locals: arguments[1] }
+            else
+              { partial: arguments[0], locals: arguments[1] }
+            end
           elsif length == 1 &&
                 arguments[0].is_a?(Prism::KeywordHashNode) &&
                 arguments[0].elements.all? do |element|
@@ -150,6 +238,7 @@ module ActionView
                 return
               end
 
+            object_template = true
             "#{dependency.pluralize}/#{dependency.singularize}"
           end